Rename file from the 'realty/services'
